feat:添加多种序列化器以及使用 Java SPI 机制实现序列化器的可定制化。

+++ example-common/src/main/java/com/xhl/example/common/service/UserService.java
@@ -16,7 +16,7 @@ public interface UserService {
     /**
      * 新方法- 获取数字
      */
-    default short getNumber() {
+    default Integer getNumber() {
 
         return 1;
     }
